Added --no-dynamic-linker flag for static linker
This flag is creates an INTERP program header entry. This entry
references the dynamic linker, which is linux specific and not
relevant for a kernel.

Removing the INTERP header is not strictly neccessary, because it
will not be interpreted by the loader in, e.g. GRUB. Still, it
should not be there.